--- amplify-tools/.gitignore
@@ -1,9 +1,0 @@
-*.iml
-.gradle
-local.properties
-.idea
-.DS_Store
-build
-captures
-.externalNativeBuild
-.cxx
