+++ amplify-tools/amplify-tools-gradle-plugin/gradle.properties
@@ -1,4 +1,4 @@
 POM_ARTIFACT_ID=amplify-tools-gradle-plugin
-POM_NAME=Amplify Tools for Android - Gradle Plugin
+POM_NAME=Amplify tools for Android - Gradle Plugin
 POM_DESCRIPTION=A Gradle plugin that installs the Amplify CLI and enables the user to execute Amplify capabilities like modelgen and push within the IDE.
 POM_PACKAGING=jar
