+++ amplify-tools/amplify-tools-gradle-plugin/src/main/resources/META-INF/gradle-plugins/com.amplifyframework.amplifytools.properties
@@ -1,1 +1,1 @@
-implementation-class=amplifytools
+implementation-class=AmplifyTools
