+++ amplify-tools/settings.gradle
@@ -1,2 +1,2 @@
 include ':amplify-tools-gradle-plugin'
-rootProject.name='amplifyTools'
+rootProject.name='AmplifyTools'
